Function to filter the mutations

This function filter the mutations in the MAF format using thresholds on various features

filterMAF2(
  maf,
  flag_genes = "default",
  save_name = NULL,
  no_filter = FALSE,
  norm_alt_max = 1,
  t_alt_min = 1,
  t_depth_min = 20,
  tumor_freq_min = 0.05,
  norm_freq_max = 0.02,
  gnomAD_AF_max = 0.001,
  AF_max = 0.001,
  ExAC_AF_max = 0.001,
  n_callers = 2,
  variant_caller = NULL
)

Arguments

maf

Dataframe object containing the mutation information in the MAF format

flag_genes

The list of genes used as flag genes

save_name

The name and path of the output file to save filtered MAFs

no_filter

Flag to filter the MAF (Default no_filter=FALSE)

norm_alt_max

Alt norm max (Default norm_alt_max=1)

t_alt_min

Alt t min (Default t_alt_min=1)

t_depth_min

Depth t min (Default t_depth_min=20)

tumor_freq_min

Tumor Frequency Minimum (Default tumor_freq_min=0.05)

norm_freq_max

norm_freq_max (Default norm_freq_max=0.02)

gnomAD_AF_max

Maximum allele frequency in gnomAD database (Default gnomAD_AF_max=0.001)

AF_max

Maximum allele frequency in 1000 genome database (Default AF_max=0.001)

ExAC_AF_max

Maximum allele frequency in ExAC database (Default ExAC_AF_max=0.01)

n_callers

Minimum number of callers identified mutation. (Default n_callers=2)

variant_caller

Name of variant caller to be used or "consensus" to apply filter based on `n_callers` (Default variant_caller=NULL)

Value

An object of class MAF with the filtered mutations

Author

Mayank Tandon, Ashish Jain

Examples

library(MAFDash)
maf <- system.file("extdata", "test.mutect2.maf.gz", package = "MAFDash")
mafDF <- read.table(maf, sep='\t', header=TRUE, fill = TRUE, quote="\"", stringsAsFactors = FALSE)
filteredMAF <- filterMAF2(maf = mafDF)