Function to filter the mutations

This function reads the the big MAF files in chunks and filter the mutations using various features

filter_maf_chunked(
  maf_file,
  chunk_lines = 10000,
  flag_genes = "default",
  save_name = NULL,
  no_filter = FALSE,
  grep_vcf_filter_col = "PASS",
  non_silent_only = FALSE,
  t_alt_min = 2,
  t_depth_min = 5,
  tumor_freq_min = 0.01,
  n_alt_max = 1,
  norm_freq_max = 0.01,
  gnomAD_AF_max = 0.001,
  AF_max = 0.001,
  ExAC_AF_max = 0.001,
  n_callers = 2,
  variant_caller = NULL
)

Arguments

maf_file
The path of the file containing the mutation information in the MAF format

chunk_lines
The number of lines to be read at once

flag_genes
The list of genes used as flag genes

save_name
The name and path of the output file to save filtered MAFs

no_filter
Flag to filter the MAF (Default no_filter=FALSE)

grep_vcf_filter_col
FILTER column

non_silent_only
Flag to filter non slient SNVs only (Default non_silent_only=FALSE)

t_alt_min
Alt t min (Default t_alt_min=1)

t_depth_min
Depth t min (Default t_depth_min=20)

tumor_freq_min
Tumor Frequency Minimum (Default tumor_freq_min=0.05)

n_alt_max
Alt norm max (Default n_alt_max=1)

norm_freq_max
norm_freq_max (Default norm_freq_max=0.02)

gnomAD_AF_max
gnomAD_AF_max (Default gnomAD_AF_max=0.001)

AF_max
AF_max (Default AF_max=0.001)

ExAC_AF_max
ExAC_AF_max (Default ExAC_AF_max=0.01)

n_callers
n_callers (Default n_callers=2)

variant_caller
variant_caller

Value

An object of class MAF with the filtered mutations

Author

Mayank Tondon, Ashish Jain

Examples

library(MAFDash)
maf <- system.file("extdata", "test.mutect2.maf.gz", package = "MAFDash")
filteredMAF <- filter_maf_chunked(maf_file = maf)