Function to generate a dashboard from a MAF file.

This function creates an HTML file containing the different figures and plots explaining the MAF dataset.

getMAFDashboard(
  maf = NULL,
  plotList = NULL,
  outputFileName = "dashboard.html",
  outputFileTitle = "MAF Dash",
  outputFilePath = tempdir()
)

Arguments

maf

MAF object or MAF file path information in the MAF format

plotList

A named list denoting which plots to draw. List elements can be: - boolean, if the name matches one of built-in plots: "summary_plot","burden","oncoplot","cooccurence","heatmap" - ggplot, ComplexHeatmap, or plotly object - file path to image (must be absolute path) The default value (if set to NULL) is 'list("summary_plot"=TRUE,"burden"=TRUE,"oncoplot"=TRUE,"cooccurence"=TRUE,"heatmap"=TRUE)' The order of the list determines the order of the tabs, and list names are used as tab titles.

outputFileName

The name of the output html dashboard file

outputFileTitle

The title of the output html dashboard

outputFilePath

The path of the output html dashboard

Value

No return value, the MAF dashboard html file is created in the given output folder

Author

Mayank Tandon, Ashish Jain

Examples

library(MAFDash)
maf <- system.file("extdata", "test.mutect2.maf.gz", package = "MAFDash")
getMAFDashboard(maf = maf,outputFilePath=tempdir())
#> 
#> 
#> processing file: MAFDash.Rmd
#>   |                                                                              |                                                                      |   0%  |                                                                              |......                                                                |   9%
#>   ordinary text without R code
#> 
#>   |                                                                              |.............                                                         |  18%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |...................                                                   |  27%
#>   ordinary text without R code
#> 
#>   |                                                                              |.........................                                             |  36%
#> label: read-MAF-file (with options) 
#> List of 1
#>  $ results: chr "hide"
#> 
#>   |                                                                              |................................                                      |  45%
#>   ordinary text without R code
#> 
#>   |                                                                              |......................................                                |  55%
#> label: make-summary-table
#>   |                                                                              |.............................................                         |  64%
#>    inline R code fragments
#> 
#>   |                                                                              |                                                                      |   0%  |                                                                              |...................................                                   |  50%
#>    inline R code fragments
#> 
#>   |                                                                              |......................................................................| 100%
#> label: make-summary-table  (with options) 
#> List of 1
#>  $ output: logi FALSE
#> 
#> 
#>   |                                                                              |...................................................                   |  73%
#> label: make-plot-tabs
#>   |                                                                              |.........................................................             |  82%
#>    inline R code fragments
#> 
#>   |                                                                              |                                                                      |   0%  |                                                                              |.......                                                               |  10%
#>   ordinary text without R code
#> 
#>   |                                                                              |..............                                                        |  20%
#> label: make-summary-maftools (with options) 
#> List of 2
#>  $ fig.width : num 12
#>  $ fig.height: num 8
#> 
#>   |                                                                              |.....................                                                 |  30%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................                                          |  40%
#> label: make-burden-plot-dot
#> Warning: `gather_()` was deprecated in tidyr 1.2.0.
#> Please use `gather()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
#>   |                                                                              |...................................                                   |  50%
#>   ordinary text without R code
#> 
#>   |                                                                              |..........................................                            |  60%
#> label: make-oncoplot (with options) 
#> List of 2
#>  $ fig.width : num 12
#>  $ fig.height: num 6
#> 
#> All mutation types: Missense Mutation, Frame Shift Del, Nonsense
#> Mutation, Splice Site, Multi Hit, In Frame Del, Frame Shift Ins.
#> `alter_fun` is assumed vectorizable. If it does not generate correct
#> plot, please set `alter_fun_is_vectorized = FALSE` in `oncoPrint()`.
#>   |                                                                              |.................................................                     |  70%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................................              |  80%
#> label: somatic-interactions (with options) 
#> List of 3
#>  $ fig.width : num 12
#>  $ fig.height: num 6
#>  $ output    : logi FALSE
#> 
#> Warning: display list redraw incomplete
#>   |                                                                              |...............................................................       |  90%
#>   ordinary text without R code
#> 
#>   |                                                                              |......................................................................| 100%
#> label: make-interactive-heatmap (with options) 
#> List of 2
#>  $ fig.width : num 12
#>  $ fig.height: num 8
#> 
#> 
#>   |                                                                              |................................................................      |  91%
#> label: make-variant-table
#>   |                                                                              |......................................................................| 100%
#>    inline R code fragments
#> 
#>   |                                                                              |                                                                      |   0%  |                                                                              |........                                                              |  11%
#>   ordinary text without R code
#> 
#>   |                                                                              |................                                                      |  22%
#> label: make-data-table (with options) 
#> List of 1
#>  $ results: chr "hide"
#> 
#> Not outputting these columns: gnomAD_AF, AF
#>   |                                                                              |.......................                                               |  33%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................                                       |  44%
#> label: set-up-variant-filters
#>   |                                                                              |.......................................                               |  56%
#>    inline R code fragments
#> 
#>   |                                                                              |                                                                      |   0%  |                                                                              |......................................................................| 100%
#> label: unnamed-chunk-2
#> 
#>   |                                                                              |...............................................                       |  67%
#> label: unnamed-chunk-1
#>   |                                                                              |......................................................                |  78%
#>   ordinary text without R code
#> 
#>   |                                                                              |..............................................................        |  89%
#> label: render-full-table
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#> 
#> output file: /private/var/folders/jy/45byld4s14q9wc9qvbwx027m7rhngt/T/Rtmp0WzLyL/file1567c4e80aa7e/MAFDash.knit.md
#> /usr/local/bin/pandoc +RTS -K512m -RTS /private/var/folders/jy/45byld4s14q9wc9qvbwx027m7rhngt/T/Rtmp0WzLyL/file1567c4e80aa7e/MAFDash.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /private/var/folders/jy/45byld4s14q9wc9qvbwx027m7rhngt/T/Rtmp0WzLyL/dashboard.html --lua-filter /Users/jaina13/Library/R/4.0/library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Users/jaina13/Library/R/4.0/library/rmarkdown/rmarkdown/lua/latex-div.lua --self-contained --variable bs3=TRUE --standalone --section-divs --template /Users/jaina13/Library/R/4.0/library/flexdashboard/rmarkdown/templates/flex_dashboard/resources/default.html --include-in-header /var/folders/jy/45byld4s14q9wc9qvbwx027m7rhngt/T//Rtmp0WzLyL/rmarkdown-str1567c61395e75.html --variable theme=yeti --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /var/folders/jy/45byld4s14q9wc9qvbwx027m7rhngt/T//Rtmp0WzLyL/rmarkdown-str1567c9aeee68.html --include-in-header /var/folders/jy/45byld4s14q9wc9qvbwx027m7rhngt/T//Rtmp0WzLyL/file1567c3ba5f2cfhtml --include-before-body /var/folders/jy/45byld4s14q9wc9qvbwx027m7rhngt/T//Rtmp0WzLyL/file1567c6dd8f35f.html --include-after-body /var/folders/jy/45byld4s14q9wc9qvbwx027m7rhngt/T//Rtmp0WzLyL/file1567c3e521a50.html --highlight-style pygments --include-before-body /var/folders/jy/45byld4s14q9wc9qvbwx027m7rhngt/T//Rtmp0WzLyL/file1567cbe812a1.html --include-after-body /var/folders/jy/45byld4s14q9wc9qvbwx027m7rhngt/T//Rtmp0WzLyL/file1567c311f0e22.html 
#> 
#> Output created: /private/var/folders/jy/45byld4s14q9wc9qvbwx027m7rhngt/T/Rtmp0WzLyL/dashboard.html